+++ src/Linode/Provider.php
@@ -29,7 +29,7 @@ use Upmind\ProvisionProviders\Servers\Data\ReinstallParams;
 use Upmind\ProvisionProviders\Servers\Data\ResizeParams;
 use Upmind\ProvisionProviders\Servers\Data\ServerIdentifierParams;
 use Upmind\ProvisionProviders\Servers\Data\ServerInfoResult;
-use Upmind\ProvisionProviders\Servers\Data\SshConnectionCommandResult;
+use Upmind\ProvisionProviders\Servers\Data\ConnectionCommandResult;
 use Upmind\ProvisionProviders\Servers\Linode\Data\Configuration;
 
 class Provider extends Category implements ProviderInterface
@@ -84,11 +84,11 @@ class Provider extends Category implements ProviderInterface
         return $this->getServerInfo($params->instance_id);
     }
 
-    public function getSshConnectionCommand(ServerIdentifierParams $params): SshConnectionCommandResult
+    public function getConnectionCommand(ServerIdentifierParams $params): ConnectionCommandResult
     {
         $server = $this->findServer((int)$params->instance_id);
 
-        return SshConnectionCommandResult::create()
+        return ConnectionCommandResult::create()
             ->setCommand(sprintf('ssh root@%s', Arr::first($server->ipv4)));
     }