+++ src/OnApp/Provider.php
@@ -21,7 +21,7 @@ use Upmind\ProvisionProviders\Servers\Data\ReinstallParams;
 use Upmind\ProvisionProviders\Servers\Data\ResizeParams;
 use Upmind\ProvisionProviders\Servers\Data\ServerIdentifierParams;
 use Upmind\ProvisionProviders\Servers\Data\ServerInfoResult;
-use Upmind\ProvisionProviders\Servers\Data\SshConnectionCommandResult;
+use Upmind\ProvisionProviders\Servers\Data\ConnectionCommandResult;
 use Upmind\ProvisionProviders\Servers\OnApp\Data\Configuration;
 
 class Provider extends Category implements ProviderInterface
@@ -81,7 +81,7 @@ class Provider extends Category implements ProviderInterface
     /**
      * @inheritDoc
      */
-    public function getSshConnectionCommand(ServerIdentifierParams $params): SshConnectionCommandResult
+    public function getConnectionCommand(ServerIdentifierParams $params): ConnectionCommandResult
     {
         try {
             $info = $this->getServerInfoResult($params->instance_id);
@@ -92,8 +92,9 @@ class Provider extends Category implements ProviderInterface
 
             $password = $this->api()->getPassword($params->instance_id);
 
-            return SshConnectionCommandResult::create()
+            return ConnectionCommandResult::create()
                 ->setMessage('SSH command generated')
+                ->setType(ConnectionCommandResult::TYPE_SSH)
                 ->setCommand(sprintf('ssh root@%s', $info['ip_address']))
                 ->setPassword($password);
         } catch (Throwable $e) {