+++ src/Virtualizor/Provider.php
@@ -15,7 +15,7 @@ use Upmind\ProvisionProviders\Servers\Data\ReinstallParams;
 use Upmind\ProvisionProviders\Servers\Data\ResizeParams;
 use Upmind\ProvisionProviders\Servers\Data\ServerIdentifierParams;
 use Upmind\ProvisionProviders\Servers\Data\ServerInfoResult;
-use Upmind\ProvisionProviders\Servers\Data\SshConnectionCommandResult;
+use Upmind\ProvisionProviders\Servers\Data\ConnectionCommandResult;
 use Upmind\ProvisionProviders\Servers\Virtualizor\Data\Configuration;
 
 /**
@@ -134,13 +134,14 @@ class Provider extends Category implements ProviderInterface
     /**
      * @inheritDoc
      */
-    public function getSshConnectionCommand(ServerIdentifierParams $params): SshConnectionCommandResult
+    public function getConnectionCommand(ServerIdentifierParams $params): ConnectionCommandResult
     {
         // $vnc = $this->api()->getVncInfo($params->instance_id);
         $info = $this->getServerInfoResult($params->instance_id);
 
-        return SshConnectionCommandResult::create()
+        return ConnectionCommandResult::create()
             ->setMessage('SSH command generated')
+            ->setType(ConnectionCommandResult::TYPE_SSH)
             ->setCommand(sprintf('ssh root@%s', $info['ip_address']));
     }